Move validate logic to _validators.py
